Add new Graph.Search function to find lowest-cost paths.
This function takes a predicate on flower.GeneticDistribution which
should return true for any acceptable distributions. It will return the
vertex of a lowest-cost path to a distribution satisfying the predicate
(or return the zero vertex & a not-OK value). The example binary is
updated to use this function rather than its own reimplementation.

This is a small change, but will enable later updates (e.g. pruning).