Refactor breedgraph.Test, include priority.
Test priority is used to determine which test takes priority if there
are multiple tests that come to the same result. This is used to ensure
that a "minimal" test is used -- e.g. if no filtering produces the same
as filtering by red, the output should always use no filtering. To
support this, breedgraph.Test was refactored to be a struct; while I was
at it, I attached the test name to the struct as well.